# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atelier Commerce API",
    "version": "1.0.0",
    "description": "Single-currency BDT commerce API. Bearer authentication; cash on delivery; server-side ownership and transactional inventory."
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders": {
      "post": {
        "summary": "Checkout with cash on delivery",
        "operationId": "postApiOrders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkout"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{8,80}$"
            },
            "description": "Reuse the same key and delivery details when retrying this checkout."
          }
        ]
      },
      "get": {
        "summary": "List own orders",
        "operationId": "getApiOrders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/orders/checkout/{key}": {
      "get": {
        "summary": "Recover a checkout confirmation by its idempotency key",
        "description": "Read-only lookup scoped to the authenticated customer. Does not place or retry an order.",
        "operationId": "getApiOrdersCheckoutKey",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Previously placed order",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "401": { "description": "Authentication required or expired" },
          "404": { "description": "No order with this key belongs to the caller" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Read own order",
        "operationId": "getApiOrdersId",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "summary": "Cancel an unshipped order; restores stock once",
        "operationId": "postApiOrdersIdCancel",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
...
    "/api/admin/orders": {
      "get": {
        "summary": "Admin: list all orders",
        "operationId": "getApiAdminOrders",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/admin/orders/{id}/status": {
      "patch": {
        "summary": "Admin: advance fulfillment or cancel before shipping",
        "operationId": "patchApiAdminOrdersIdStatus",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Status"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Checkout": {
        "type": "object",
        "properties": {
          "recipient": {
            "type": "string",
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "pattern": "^[+0-9 ()-]{7,30}$"
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "city": {
            "type": "string",
            "maxLength": 100
          },
          "expectedTotal": {
            "type": "number",
            "minimum": 0.01,
            "description": "Total displayed in the reviewed cart. Checkout returns 409 if recalculated total differs."
          }
        },
        "required": [
          "recipient",
          "phone",
          "address",
          "city",
          "expectedTotal"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH_ON_DELIVERY"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "UNPAID",
              "PAID",
              "CANCELLED"
            ]
          },
          "recipient": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "shipping": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "OrderPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "page": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer"
          }
        }
      }
    }
  }
}